% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertisement.R
\docType{data}
\name{advertisement}
\alias{advertisement}
\title{Social Network Ads}
\format{A data frame with 400 rows and 5 variables}
\usage{
advertisement
}
\description{
This is an example data used in logistic regression
}
\details{
This data is from Kaggle (https://www.kaggle.com/rakeshrau/social-network-ads)
}
\examples{
data(advertisement)
data("advertisement")
}
\keyword{datasets}
