% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloncancer.R
\docType{data}
\name{coloncancer}
\alias{coloncancer}
\title{Colon Cancer Gene Expression Data}
\format{A data frame with 62 rows and 2001 variables}
\usage{
coloncancer
}
\description{
This is an example data used in MFSpcapls function
}
\details{
This data is from Kaggle (https://www.kaggle.com/masudur/colon-cancer-gene-expression-data)
}
\examples{
data(coloncancer)
data("coloncancer")
}
\keyword{datasets}
