% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Medical Cost Personal Datasets}
\format{A data frame with 1338 rows and 7 variables}
\usage{
insurance
}
\description{
This is an example data used in linear regression
}
\details{
This data is from Kaggle (https://www.kaggle.com/mirichoi0218/insurance)
}
\examples{
data(insurance)
data("insurance")
}
\keyword{datasets}
