% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung.R
\docType{data}
\name{lung}
\alias{lung}
\title{Medical Cost Personal Datasets}
\format{A data frame with 137 rows and 9 variables}
\usage{
lung
}
\description{
This is an example data used in linear regression
}
\details{
This data is from Vanderbilt Biostatistics Wiki (http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/Cvalung.html)
}
\examples{
data(lung)
data("lung")
}
\keyword{datasets}
