% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mephasOpen.R
\name{mephasOpen}
\alias{mephasOpen}
\title{MEPHAS Open Windows}
\usage{
mephasOpen(
  method = c("condist", "disdist", "ttest", "nptest", "proptest", "tabletest", "anova",
    "linereg", "logisreg", "surv", "pca", "pls")
)
}
\arguments{
\item{method}{choose method from MEPHAS}
}
\value{
shiny interface
}
\description{
MEPHAS open function
}
\examples{
# mephas::mephasOpen("condist")
## or,
# library(mephas)
# mephasOpen("condist")
}
