% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_btest.R
\name{mfs_btest}
\alias{mfs_btest}
\title{Tests based on binomial proportional data}
\usage{
mfs_btest()
}
\value{
The shiny web page of the tests based on binominal proportion
}
\description{
This is the description of this function, 3
}
\details{
This is the details of this function, 3
}
\examples{
mfs_btest()
}
