% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_crotb.R
\name{mfs_crotb}
\alias{mfs_crotb}
\title{Tests for cross tab data}
\usage{
mfs_crotb()
}
\value{
The shiny web page of the tests for cross tab data
}
\description{
This is the description of this function, 4
}
\details{
This is the details of this function, 4
}
\examples{
mfs_crotb()
}
