% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_dist.R
\name{mfs_dist}
\alias{mfs_dist}
\title{Tests for cross tab data}
\usage{
mfs_dist()
}
\value{
The shiny web page of the tests for cross tab data
}
\description{
This is the description of this function, 5
}
\details{
This is the details of this function, 4
}
\examples{
mfs_dist()
}
