% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_nptest.R
\name{mfs_nptest}
\alias{mfs_nptest}
\title{Non-paramatric tests}
\usage{
mfs_nptest()
}
\value{
The shiny web page of non-parametric statistical test
}
\description{
This is the description of this function
}
\details{
This is the details of this function
}
\examples{
mfs_nptest()
}
