% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_ttest.R
\name{mfs_ttest}
\alias{mfs_ttest}
\title{T tests on one-sample and two samples}
\usage{
mfs_ttest()
}
\value{
The shiny web page of T test for one sample and two samples
}
\description{
This is the description of this function
}
\details{
This is the details of this function
}
\examples{
mfs_ttest()
}
