% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_ureg.R
\name{mfs_ureg}
\alias{mfs_ureg}
\title{Tests for cross tab data}
\usage{
mfs_ureg()
}
\value{
The shiny web page of the tests for cross tab data
}
\description{
This is the description of this function, 4
}
\details{
This is the details of this function, 4
}
\examples{
mfs_ureg()
}
