% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSanova.R
\name{MFSanova}
\alias{MFSanova}
\title{MEPHAS Shiny Application of ANOVA and Multiple Comparison}
\usage{
MFSanova()
}
\value{
The shiny app and web page of the ANOVA and multiple comparison
}
\description{
MFSanova function creates a dynamic calculator which enables users to do ANOVA analysis and multiple comparison. Users can either input data manually or upload their dataset.
}
\details{
This app includes one-way ANOVA, two-way ANOVA, and multiple comparison. The results include ANOVA table, descriptive statistics, and bar-plot of the group's mean value.
Please click "close" window to quit the application. "Rmisc" and "psych" packages are required.
}
\examples{
# library(mephas)
# MFSanova()
}
