% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFSbox1}
\alias{MFSbox1}
\title{plot functions in MEPHAS}
\usage{
MFSbox1(data, varx)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}
}
\description{
plot functions in MEPHAS
}
