% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFScondist.R
\name{MFScondist}
\alias{MFScondist}
\title{MEPHAS: Continuous Probability Distribution (Probability)}
\usage{
MFScondist()
}
\value{
shiny interface
}
\description{
MFScondist includes probability distribution of
(1) normal distribution
(2) exponential distribution
(3) gamma distribution
(4) beta distribution
(5) t distribution
(6) chi-square distribution
and (7) F distribution.
}
\details{
MFScondist also generates random numbers draw the distribution of user data
}
\examples{
# mephas::MFScondist()
## or,
# library(mephas)
# MFScondist()
}
