% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSdistribution.R
\name{MFSdistribution}
\alias{MFSdistribution}
\title{MEPHAS Shiny application of the Statistical Distributions}
\usage{
MFSdistribution()
}
\value{
The shiny web page of the statistical distribution
}
\description{
MFSdistribution function creates a dynamic calculator which shows the commonly used probability distribution. Users can change the parameter to change the shape of distribution.
}
\details{
This app includes the distributions for continuous variables (normal distribution, exponential distribution, and gamma distribution), for discrete variables (binomial distribution and Poisson distribution), and distributions derived from normal distribution (t distribution, chi-square distribution, and F distribution).
Please click "close" window to quit the application. No special packages are required.
}
\examples{
# library(mephas)
# MFSdist()
}
