% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFShist1c}
\alias{MFShist1c}
\title{plot functions in MEPHAS}
\usage{
MFShist1c(data, varx, bw)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}

\item{bw}{input bin numbers}
}
\description{
plot functions in MEPHAS
}
