% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFShist2}
\alias{MFShist2}
\title{plot functions in MEPHAS}
\usage{
MFShist2(data2, bw)
}
\arguments{
\item{data2}{input data frame}

\item{bw}{input bin numbers}
}
\description{
plot functions in MEPHAS
}
