% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFSload}
\alias{MFSload}
\title{plot functions in MEPHAS}
\usage{
MFSload(loads, a)
}
\arguments{
\item{loads}{input loading data frame}

\item{a}{number of components}
}
\description{
plot functions in MEPHAS
}
