% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSlogit.R
\name{MFSlogit}
\alias{MFSlogit}
\title{MEPHAS: Logistic Regression Model (Advanced Method)}
\usage{
MFSlogit()
}
\value{
shiny interface
}
\description{
MFSlogit includes
(1) logistic regression for binary outcomes,
}
\examples{
# library(mephas)
# MFSlogit()
# or,
# mephas::MFSlogit()
# or,
# mephasOpen("logisreg")
# Use 'Stop and Quit' Button in the top to quit the interface
}
