% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSlr.R
\name{MFSlr}
\alias{MFSlr}
\title{MEPHAS: Linear Regression Model (Advanced Method)}
\usage{
MFSlr()
}
\value{
shiny interface
}
\description{
MFSlr includes
(1) linear regression for continuous outcomes
}
\examples{
# library(mephas)
# MFSlr()
# or,
# mephas::MFSlr()
# or,
# mephasOpen("linereg")
# Use 'Stop and Quit' Button in the top to quit the interface
}
