% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFSmsdm}
\alias{MFSmsdm}
\title{plot functions in MEPHAS}
\usage{
MFSmsdm(datam, var, grp)
}
\arguments{
\item{datam}{input data frame}

\item{var}{input variable name}

\item{grp}{input factor group variable name}
}
\description{
plot functions in MEPHAS
}
