% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSnptest.R
\name{MFSnptest}
\alias{MFSnptest}
\title{MEPHAS Shiny Application of the Non-paramatric Tests}
\usage{
MFSnptest()
}
\value{
The shiny web page of non-parametric statistical test
}
\description{
MFSnptest function creates a dynamic calculator which enables users to do non-parametric statistical tests. Users can either input data manually or upload their dataset.
}
\details{
This app includes non-parametric tests for one sample, two samples, and two paired samples. The results include box-plots, histogram, descriptive statistics,and test results.
Please click "close" window to quit the application. "gridExtra", "reshape", "pastecs", "DescTools", and "RVAideMemoire" packages may be required.
}
\examples{
# library(mephas)
# MFSn()
# not run
}
