% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSpca.R
\name{MFSpca}
\alias{MFSpca}
\title{MEPHAS: Dimensional analysis 1 (Advanced Method)}
\usage{
MFSpca()
}
\value{
shiny interface
}
\description{
MFSpca includes
(1) principal component analysis
and (2) factor analysis
}
\examples{
if (interactive()) {
 MFSpca()
}

if (interactive()) {
 mephasOpen("pca")
}

}
