% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSpcapls.R
\name{MFSpcapls}
\alias{MFSpcapls}
\title{MEPHAS Shiny Application of PCA PLS Regression}
\usage{
MFSpcapls()
}
\value{
The shiny web page of the PCA PLS regression
}
\description{
MFSpcapls function creates a dynamic calculator which enables users to do PCA, PLS, sparse PLS regression. Users can either input data manually or upload their dataset.
}
\details{
This app includes one-way ANOVA, two-way ANOVA, and multiple comparison. The results include ANOVA table, descriptive statistics, and bar-plot of the group's mean value.
Please click "close" window to quit the application. "ggfortify", spls" and "pls" package may be required.
}
\examples{
# library(mephas)
# MFSpcapls()
# not run
}
