% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSpls.R
\name{MFSpls}
\alias{MFSpls}
\title{MEPHAS: Dimensional analysis 2 (Advanced Method)}
\usage{
MFSpls()
}
\value{
The web-based GUI and interactive interfaces
}
\description{
MFSpls includes
(1) principal component regression
(2) partial least squares regression
and (3) sparse partial least squares regression
}
\examples{
# library(mephas)
# MFSpls()
# or,
# mephas::MFSpls()
# or,
# mephasOpen("pls")
# Use 'Stop and Quit' Button in the top to quit the interface
}
