% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSproptest.R
\name{MFSproptest}
\alias{MFSproptest}
\title{MEPHAS Shiny application of the binomial proportional data tests}
\usage{
MFSproptest()
}
\value{
The shiny web page of the tests based on binomial proportion
}
\description{
MFSproptest function creates a dynamic calculator which enables users to do statistical tests on binomial proportional data. Users can either input data manually or upload their dataset.
}
\details{
This app includes statistical tests for one single proportion, two proportions from independent groups, and two proportions from the matched data. Results include the corresponding test's outcome and pie-plot of data.
Please click "close" window to quit the application. "gridExtra", "reshape", and "pastecs" packages are required.
}
\examples{
# library(mephas)
# MFSproptest()
# not run
}
