% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSrctabtest.R
\name{MFSrctabtest}
\alias{MFSrctabtest}
\title{MEPHAS Shiny Application of Contingency Table Related Tests}
\usage{
MFSrctabtest()
}
\value{
The shiny web page of the tests for cross tab data
}
\description{
MFSrctabtest function creates a dynamic calculator which enables users to do statistical tests on R by C data. Users can either input data manually or upload their dataset.
}
\details{
This app includes statistical tests for R by C table, trend test for 2 by K table, and goodness-of-fit test.
Please click "close" window to quit the application. "psych" package is required.
}
\examples{
# library(mephas)
# MFSrctabtest()
# not run
}
