% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSreg.R
\name{MFSreg}
\alias{MFSreg}
\title{MEPHAS Shiny Application of Univariate Regression}
\usage{
MFSreg()
}
\value{
The shiny web page of the tests univariate regression
}
\description{
MFSreg function creates a dynamic calculator which instructs users to do linear regression, logistic regression, and cox regression. Users can either input data manually or upload their dataset.
}
\details{
This app includes linear regression, logistic regression, and cox regression. The results include ANOVA table, descriptive statistics, and residual plots.
Please click "close" window to quit the application. "survival", "survminer", and "ggfortify" packages may be required.
}
\examples{
# library(mephas)
# MFSreg()
# not run
}
