% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFSscorec}
\alias{MFSscorec}
\title{plot functions in MEPHAS}
\usage{
MFSscorec(scores, n1, n2, type)
}
\arguments{
\item{scores}{input score data frame}

\item{n1}{input nth component for x-axis}

\item{n2}{input nth component for y-axis}

\item{type}{the type of circle}
}
\description{
plot functions in MEPHAS
}
