% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{MFSslgt}
\alias{MFSslgt}
\title{plot functions in MEPHAS}
\usage{
MFSslgt(data, varx, vary)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}

\item{vary}{input y variable name}
}
\description{
plot functions in MEPHAS
}
