% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSsurv.R
\name{MFSsurv}
\alias{MFSsurv}
\title{MEPHAS: Survival Analysis (Advanced Method)}
\usage{
MFSsurv()
}
\value{
shiny interface
}
\description{
MFSsurv includes
(1) K-M estimation and log-rank test for survival probability curves,
(2) Cox regression
and (3) accelerated failure time model
}
\examples{
#mephas::MFSsurv()
## or,
# library(mephas)
# MFSsurv()
}
