% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSttest.R
\name{MFSttest}
\alias{MFSttest}
\title{MEPHAS Shiny Application of T test.}
\usage{
MFSttest()
}
\value{
The shiny application web page of T test for one sample and two samples
}
\description{
MFSttest function creates a dynamic calculator which enables users to do t tests . Users can either input data manually or upload their dataset.
}
\details{
This app includes t tests for one sample, two samples, and two paired samples. The results include box-plots, histogram, descriptive statistics,and test results. "gridExtra", "reshape", "pastecs" package may be required.
}
\examples{
# library(mephas)
# MFSttest()
# not run
}
