% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nptest.p3.ui.R
\name{nptest.wstest.p}
\alias{nptest.wstest.p}
\title{UI of Wilcoxon Signed-Rank Test (Non-paramatric Tests)}
\usage{
nptest.wstest.p()
}
\description{
UI of Wilcoxon Signed-Rank Test (Non-paramatric Tests)
}
