% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{plot_3D}
\alias{plot_3D}
\title{plot functions in MEPHAS}
\usage{
plot_3D(scores, loads, nx, ny, nz, scale)
}
\arguments{
\item{scores}{input score data frame}

\item{loads}{input loading data frame}

\item{nx}{input nth component for x-axis}

\item{ny}{input nth component for y-axis}

\item{nz}{input nth component for x-axis}

\item{scale}{input length}
}
\description{
plot functions in MEPHAS
}
