% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_density1}
\alias{plot_density1}
\title{plot functions in MEPHAS}
\usage{
plot_density1(data, varx)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}
}
\description{
plot functions in MEPHAS
}
