% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_hist1}
\alias{plot_hist1}
\title{plot functions in MEPHAS}
\usage{
plot_hist1(data, varx, bw)
}
\arguments{
\item{data}{input data frame}

\item{varx}{input x variable name}

\item{bw}{input bin numbers}
}
\description{
plot functions in MEPHAS
}
