% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_hist2}
\alias{plot_hist2}
\title{plot functions in MEPHAS}
\usage{
plot_hist2(data2, bw)
}
\arguments{
\item{data2}{input data frame}

\item{bw}{input bin numbers}
}
\description{
plot functions in MEPHAS
}
