% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_line2}
\alias{plot_line2}
\title{plot functions in MEPHAS}
\usage{
plot_line2(data2, var, grp1, grp2)
}
\arguments{
\item{data2}{input data frame}

\item{var}{input variable name}

\item{grp1}{input factor group 1 name}

\item{grp2}{input factor group 2 name}
}
\description{
plot functions in MEPHAS
}
