% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_linely}
\alias{plot_linely}
\title{plot functions in MEPHAS}
\usage{
plot_linely(data, vary, vx1, vx2)
}
\arguments{
\item{data}{input data frame}

\item{vary}{input y variable name}

\item{vx1}{input x1 variable name}

\item{vx2}{input x2 variable name}
}
\description{
plot functions in MEPHAS
}
