% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{plot_mat}
\alias{plot_mat}
\title{plot functions in MEPHAS}
\usage{
plot_mat(data, ybreak)
}
\arguments{
\item{data}{input data frame}

\item{ybreak}{input variable name as y-break}
}
\description{
plot functions in MEPHAS
}
