% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_msdm}
\alias{plot_msdm}
\title{plot functions in MEPHAS}
\usage{
plot_msdm(datam, var, grp)
}
\arguments{
\item{datam}{input data frame}

\item{var}{input variable name}

\item{grp}{input factor group variable name}
}
\description{
plot functions in MEPHAS
}
