% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_scorec}
\alias{plot_scorec}
\title{plot functions in MEPHAS}
\usage{
plot_scorec(scores, n1, n2, groupvec, type)
}
\arguments{
\item{scores}{input score data frame}

\item{groupvec}{input group variable}

\item{type}{input type variable}

\item{varx}{input x variable}

\item{vary}{input y variable}
}
\description{
plot functions in MEPHAS
}
