% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_scoreg}
\alias{plot_scoreg}
\title{plot functions in MEPHAS}
\usage{
plot_scoreg(scores, n1, n2, groupvec)
}
\arguments{
\item{scores}{input score data frame}

\item{n1}{input x variable}

\item{n2}{input y variable}

\item{groupvec}{input group variable}
}
\description{
plot functions in MEPHAS
}
