% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{tabPanel.upload.num}
\alias{tabPanel.upload.num}
\title{panel functions in MEPHAS}
\usage{
tabPanel.upload.num(file = "file", header = "header", col = "col", sep = "sep")
}
\arguments{
\item{file}{input data}

\item{header}{input header name}

\item{col}{input column  name}

\item{sep}{input separator name}
}
\description{
panel functions in MEPHAS
}
