% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mephasOpen.R
\name{mephasOpen}
\alias{mephasOpen}
\title{Open MEPHAS  Windows}
\usage{
mephasOpen(
  method = c("condist", "disdist", "ttest", "nptest", "proptest", "tabletest", "anova",
    "linereg", "logisreg", "surv", "pca", "pls")
)
}
\arguments{
\item{method}{choose method from MEPHAS,
"condist" opens continuous distribution,
"disdist" opens discrete distribution,
"ttest" opens t test,
"nptest" opens non-parametric test,
"proptest" opens test methods for proportions,
"tabletest" opens test methods contingency table,
"anova"opens ANOVA,
"linereg" opens linear regression,
"logisreg" opens logistic regression,
"surv" opens survival models,
"pca" opens principal component analysis,
"pls" opens partial least square methods}
}
\value{
shiny interface of MEPHAS (see web-based version: https://alain003.phs.osaka-u.ac.jp/mephas/)
}
\description{
Open MEPHAS interfaces
}
\examples{
if (interactive()) {
  mephasOpen("condist")
}


}
