% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSanova.R
\name{MFSanova}
\alias{MFSanova}
\title{MEPHAS: ANOVA (Hypothesis Testing)}
\usage{
MFSanova()
}
\value{
shiny interface
}
\description{
MFSanova includes
(1) one-way ANOVA,
(2) pairwise post-hoc test for one-way ANOVA,
(3) two-way ANOVA,
(4) pairwise post-hoc test for two-way ANOVA
(5) Kruskal-Wallis test
and (6) post-hoc test for Kruskal-Wallis test
}
\examples{
if (interactive()) {
 MFSanova()
}

if (interactive()) {
 mephasOpen("anova")
}
}
