% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSdisdist.R
\name{MFSdisdist}
\alias{MFSdisdist}
\title{MEPHAS: Discrete Probability Distribution (Probability)}
\usage{
MFSdisdist()
}
\value{
shiny interface
}
\description{
MFSdisdist includes probability distribution of
(1) binomial distribution
and (2) Poisson distribution
}
\details{
MFScondist also generates random numbers draw the distribution of user data
}
\examples{
# mephas::MFSdisdist()
## or,
# library(mephas)
# MFSdisdist()
}
