% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSlr.R
\name{MFSlr}
\alias{MFSlr}
\title{MEPHAS: Linear Regression Model (Advanced Method)}
\usage{
MFSlr()
}
\value{
shiny interface
}
\description{
MFSlr includes
(1) linear regression for continuous outcomes
}
\examples{
if (interactive()) {
 MFSlr()
}

if (interactive()) {
 mephasOpen("linereg")
}

}
