% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSnptest.R
\name{MFSnptest}
\alias{MFSnptest}
\title{MEPHAS: Non-parametric Test (Hypothesis Testing)}
\usage{
MFSnptest()
}
\value{
shiny interface
}
\description{
MFSnptest includes Non-parametric test of
(1) one sample,
(2) two independent samples,
and (3) two paried samples.
}
\examples{
if (interactive()) {
 MFSnptest()
}

if (interactive()) {
 mephasOpen("nptest")
}


}
