% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSproptest.R
\name{MFSproptest}
\alias{MFSproptest}
\title{MEPHAS: Test for Binomial Proportion (Hypothesis Testing)}
\usage{
MFSproptest()
}
\value{
shiny interface
}
\description{
MFSproptest includes test for binomial proportion of
(1) one single proportion,
(2) two proportions from independent groups,
(3) more than two groups,
and (4) trend in more than two groups.
}
\examples{
if (interactive()) {
 MFSproptest()
}

if (interactive()) {
 mephasOpen("proptest")
}
}
