% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFSttest.R
\name{MFSttest}
\alias{MFSttest}
\title{MEPHAS: T Test (Hypothesis Testing)}
\usage{
MFSttest()
}
\value{
shiny interface
}
\description{
MFSttest includes t test of
(1) one sample,
(2) two independent samples,
and (3) two paired samples.
}
\examples{
# library(mephas)
# MFSttest()
# or,
# mephas::MFSttest()
# or,
# mephasOpen("ttest")
# Use 'Stop and Quit' Button in the top to quit the interface
}
