% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_3D}
\alias{plot_3D}
\title{plot functions in MEPHAS}
\usage{
plot_3D(scores, loads, nx, ny, nz, scale)
}
\arguments{
\item{scores}{input score data}

\item{loads}{input loading data}

\item{nx}{input x1 variable name}

\item{ny}{input x2 variable name}

\item{nz}{input x3 variable name}

\item{scale}{input line length}
}
\description{
plot functions in MEPHAS
}
