% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{plot_biplot}
\alias{plot_biplot}
\title{plot functions in MEPHAS}
\usage{
plot_biplot(scores, loads, n1, n2)
}
\arguments{
\item{scores}{input score data frame}

\item{loads}{input loading data frame}

\item{n1}{input nth component for x-axis}

\item{n2}{input nth component for y-axis}
}
\description{
plot functions in MEPHAS
}
