% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_score}
\alias{plot_score}
\title{plot functions in MEPHAS}
\usage{
plot_score(scores, n1, n2)
}
\arguments{
\item{scores}{input score data frame}

\item{n1}{input x variable}

\item{n2}{input y variable}
}
\description{
plot functions in MEPHAS
}
